Move IntensityTable overlap code and tests inside IntensityTable package
IntensityTable overlap code not going to be reused for ImageStack, so might as well keep it together.

Test plan: `pytest -v -n4 starfish/intensity_table/`

Depends on #1169
